Rename module `text` -> `texty`.
