+++ src/loggy-intf/package.json
@@ -18,7 +18,7 @@
     "@this/async": "*",
     "@this/decon": "*",
     "@this/quant": "*",
-    "@this/text": "*",
+    "@this/texty": "*",
     "@this/typey": "*",
     "@this/valvis": "*"
   }
