+++ src/texty/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/typey",
+  "name": "@this/texty",
   "version": "0.0.1",
   "type": "module",
   "private": true,
